requirements.txt
aiogram==3.4.1
python-dotenv
requests
beautifulsoup4
